fish

+++ config/dot_config/fish/conf.d/4-inshellisense.fish
@@ -1,5 +1,5 @@
 # ---------------- inshellisense shell plugin ----------------
-if test -z "$ISTERM" && status --is-interactive
+if test -z "$ISTERM" && status --is-interactive && type -q is
   if status --is-login
     is -s fish --login ; kill %self
   else
